Refactored away authentication stuff to Authenticator class
