+++ .gitignore
@@ -1,3 +1,4 @@
 phpunit.xml
 composer.lock
 vendor
+index.php
